Fixed an issue with auto-scroll in Library window when there is no area to scroll Fixed GUI tree view so that selection overlay is removed when mouse leaves the area Potentially fixing an issue where undocking a window was causing a crash because GUIWidget was alive for a frame longer than its parent window

+++ MBansheeEditor/Inspector/InspectableFloat.cs
@@ -47,7 +47,7 @@ namespace BansheeEditor
         public override InspectableState Refresh(int layoutIndex)
         {
             if (guiFloatField != null && !guiFloatField.HasInputFocus)
-                guiFloatField.Value = property.GetValue<int>();
+                guiFloatField.Value = property.GetValue<float>();
 
             InspectableState oldState = state;
             if (state.HasFlag(InspectableState.Modified))
